+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- snapshot_dfn_modify() and properties to responses.SnapshotDefinition
+
 ### Fixed
 
 - Error out if reading controller version fails